upd and fix

+++ memoryrange.cpp
@@ -45,8 +45,8 @@ void Memory_range_type::Allocated(const bool &VAllocated) {
 bool Memory_range_type::Allocate(const Weighted_size &Relative_end) {
   auto Result = false;
   if (!Allocated_or_adopted() && Relative_end.Valid()) {
-    Range().Start() = reinterpret_cast<size_t>(malloc(Relative_end + 1));
-    Range().Start().Weight() = Range().Start().Value() != 0;
+    Range().Start() = reinterpret_cast<size_t>(malloc(Relative_end + static_cast<size_t>(1)));
+    Range().Start().Weight() = Range().Start().Value() != static_cast<size_t>(0);
     Range().End_from_relative_end(Relative_end);
     Result = Range().Start().Weight();
     Allocated(Result);
@@ -59,8 +59,8 @@ bool Memory_range_type::Reallocate(Weighted_size Relative_end) {
   if (Allocated() && Relative_end.Valid() &&
       Relative_end != Range().Relative_end()) {
     Range().Start() = reinterpret_cast<size_t>(realloc(
-        reinterpret_cast<void *>(Range().Start().Value()), Relative_end + 1));
-    Range().Start().Weight() = Range().Start().Value() != 0;
+        reinterpret_cast<void *>(Range().Start().Value()), Relative_end + static_cast<size_t>(1)));
+    Range().Start().Weight() = Range().Start().Value() != static_cast<size_t>(0);
     Range().End_from_relative_end(Relative_end);
     Result = Range().Start().Weight();
     Allocated(Result);
@@ -112,10 +112,10 @@ const Range_of_size &Memory_range_type::Range() const { return _Range; }
 
 template <class TemplateClass>
 bool _MemoryWipe_with_zeroBufferRaw(size_t &Start, size_t End) {
-  if ((Start <= End) && (End - Start >= (sizeof(TemplateClass) - 1))) {
+  if ((Start <= End) && (End - Start >= (sizeof(TemplateClass) - static_cast<size_t>(1)))) {
     const TemplateClass Buffer = 0;
     auto CurrentStart = Start;
-    const auto StartMax = End - (sizeof(TemplateClass) - 1);
+    const auto StartMax = End - (sizeof(TemplateClass) - static_cast<size_t>(1));
     while (CurrentStart <= StartMax) {
       *reinterpret_cast<TemplateClass *>(CurrentStart) = Buffer;
       if (CurrentStart < StartMax) {
@@ -204,7 +204,7 @@ void _MemoryShiftUpWithBufferRaw(size_t &CurrentSourceStart, size_t SourceEnd,
 void _MemoryShiftUpRaw(size_t SourceStart, size_t SourceEnd,
                        size_t DestinationStart) {
   auto StepForward = DestinationStart - SourceStart;
-  auto StepBack = SourceEnd - SourceStart + 1;
+  auto StepBack = SourceEnd - SourceStart + static_cast<size_t>(1);
 
   auto BufferSizeLeft = GCD(StepForward, StepBack);
 
@@ -246,8 +246,8 @@ bool Memory_range_type::Shift_exchange(Range_of_size Source,
       _MemoryShiftUpRaw(Source.Start(), Source.End(), Destination.Start());
     } else {
       // Convert shift down to shift up
-      _MemoryShiftUpRaw(Destination.Start(), Source.Start() - 1,
-                        Destination.End() + 1);
+      _MemoryShiftUpRaw(Destination.Start(), Source.Start() - static_cast<size_t>(1),
+                        Destination.End() + static_cast<size_t>(1));
     }
   }
   return Result;
